backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.1
astroid==3.3.5
autopep8==2.0.2
billiard==4.2.1
blinker==1.9.0
braintree==4.20.0
cachetools==5.5.0
celery==5.4.0
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
dill==0.3.9
distlib==0.4.0
Django==4.2.7
django-colorfield==0.10.1
django-cors-headers==4.3.1
django-crequest==2018.5.11
django-extra-fields==3.0.2
django-model-utils==4.3.1
django-modeltranslation==0.19.16
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-rest-framework-docs==0.1.7
django-solo==2.4.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-yasg==1.21.5
factory-boy==3.2.1
Faker==18.3.2
filelock==3.18.0
Flask==3.1.2
flask-cors==6.0.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.36.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
httplib2==0.31.0
idna==3.4
inflection==0.5.1
isort==5.13.2
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.2
jsonfield==3.1.0
jsonpickle==3.0.1
kombu==5.4.2
MarkupSafe==2.1.2
mccabe==0.7.0
mysqlclient==2.2.0
packaging==23.0
paho-mqtt==1.6.1
pdf2image==1.17.0
pdfminer.six==20251107
pdfplumber==0.11.8
phonenumbers==8.13.8
Pillow==9.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.6.0
pylint==3.3.1
pyOpenSSL==23.1.1
pyparsing==3.2.5
pypdfium2==5.0.0
pytesseract==0.3.13
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3
redis==5.2.0
requests==2.28.2
rsa==4.9
ruamel.yaml==0.18.11
ruamel.yaml.clib==0.2.12
sendgrid==6.10.0
six==1.16.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
stripe==12.3.0
tomli==2.0.1
tomlkit==0.13.2
tqdm==4.67.1
twilio==7.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.15
vine==5.1.0
virtualenv==20.33.0
wcwidth==0.2.13
Werkzeug==3.1.3
